requirements.txt
bs4
pillow
PySocks
psutil
requests
py-cord
cairosvg
